refactor and update docs
