+++ src/components/Info.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 
-import Banner from './Banner';
+import { Banner } from '../components';
 
 const InfoIcon = props => {
   return (
